+++ Cargo.toml
@@ -10,7 +10,7 @@ members = [
     "core/threshold-experiments",
     "core/threshold-algebra",
     "core/threshold-execution",
-    "core/threshold-experimental",
+    "core/threshold-bgv",
     "core/threshold-hashing",
     "core/threshold-networking",
     "core/threshold-types",
@@ -81,7 +81,7 @@ backward-compatibility = { path = "./backward-compatibility" }  # Internal backw
 bc2wrap = { path = "./bc2wrap" }  # Internal bincode wrapper custom implementation
 error-utils = { path = "./core/error-utils", default-features = false }
 threshold-execution = { path = "./core/threshold-execution", default-features = false }
-threshold-experimental = { path = "./core/threshold-experimental", default-features = false }
+threshold-bgv = { path = "./core/threshold-bgv", default-features = false }
 hashing = { package = "threshold-hashing", path = "./core/threshold-hashing", default-features = false }
 kms = { path = "./core/service", default-features = false }  # Core KMS service
 kms-core-client = { path = "./core-client" }  # Core KMS client library
