+++ core/threshold-bgv/src/gen_bits_odd.rs
