+++ core/threshold-experiments/Cargo.toml
@@ -49,7 +49,7 @@ tonic = { workspace = true, features = ["tls-ring"] }
 tonic-health.workspace = true
 tower = { workspace = true, features = ["retry", "timeout"] }
 tower-http = { workspace = true, features = ["trace"] }
-threshold-experimental.workspace = true
+threshold-bgv.workspace = true
 threshold-types.workspace = true
 tracing = { workspace = true, features = ["log"] }
 x509-parser.workspace = true
@@ -61,7 +61,7 @@ x509-parser.workspace = true
 criterion = { workspace = true, features = ["async_tokio", "html_reports"] }
 crypto-bigint.workspace = true
 threshold-execution = { workspace = true, features = ["testing"] }
-threshold-experimental.workspace = true
+threshold-bgv.workspace = true
 paste.workspace = true
 pprof.workspace = true
 rayon.workspace = true
