+++ core/threshold-experiments/benches/algebra.rs
@@ -14,7 +14,7 @@ use rand::SeedableRng;
 fn bench_lagrange_poly(c: &mut Criterion) {
     // params are (num_parties, threshold, max_errors)
 
-    use threshold_experimental::algebra::levels::LevelOne;
+    use threshold_bgv::algebra::levels::LevelOne;
     let params = vec![(4, 1, 0), (10, 3, 0), (10, 3, 2), (40, 13, 0)];
     let mut group = c.benchmark_group("lagrange_interpolation");
