+++ core/threshold-experiments/benches/non-threshold/bgv/memory/basic_ops.rs
@@ -8,7 +8,7 @@ mod utilities;
 
 use rand::RngCore;
 use tfhe::core_crypto::seeders::new_seeder;
-use threshold_experimental::{
+use threshold_bgv::{
     algebra::{
         levels::{LevelEll, LevelKsw},
         ntt::*,
