+++ core/threshold-experiments/src/choreography/bgv/grpc.rs
@@ -29,17 +29,17 @@ use threshold_execution::runtime::sessions::small_session::SmallSession;
 use threshold_execution::small_execution::prss::{
     DerivePRSSState, PRSSInit, PRSSSetup, RobustSecurePrssInit,
 };
-use threshold_experimental::algebra::levels::{LevelEll, LevelKsw, LevelOne};
-use threshold_experimental::algebra::ntt::{Const, N65536};
-use threshold_experimental::bgv::basics::{PrivateBgvKeySet, PublicBgvKeySet, PublicKey};
-use threshold_experimental::bgv::ddec::noise_flood_decryption;
-use threshold_experimental::bgv::dkg::bgv_distributed_keygen;
-use threshold_experimental::bgv::dkg_orchestrator::BGVPreprocessingOrchestrator;
-use threshold_experimental::bgv::dkg_preproc::InMemoryBGVDkgPreprocessing;
-use threshold_experimental::bgv::utils::transfer_secret_key;
-use threshold_experimental::bgv::utils::{gen_key_set, transfer_pub_key};
-use threshold_experimental::constants::INPUT_PARTY_ID;
-use threshold_experimental::constants::PLAINTEXT_MODULUS;
+use threshold_bgv::algebra::levels::{LevelEll, LevelKsw, LevelOne};
+use threshold_bgv::algebra::ntt::{Const, N65536};
+use threshold_bgv::bgv::basics::{PrivateBgvKeySet, PublicBgvKeySet, PublicKey};
+use threshold_bgv::bgv::ddec::noise_flood_decryption;
+use threshold_bgv::bgv::dkg::bgv_distributed_keygen;
+use threshold_bgv::bgv::dkg_orchestrator::BGVPreprocessingOrchestrator;
+use threshold_bgv::bgv::dkg_preproc::InMemoryBGVDkgPreprocessing;
+use threshold_bgv::bgv::utils::transfer_secret_key;
+use threshold_bgv::bgv::utils::{gen_key_set, transfer_pub_key};
+use threshold_bgv::constants::INPUT_PARTY_ID;
+use threshold_bgv::constants::PLAINTEXT_MODULUS;
 use threshold_networking::choreography_gen::choreography_server::{
     Choreography, ChoreographyServer,
 };
